Address Doron's comments
